# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: SIFT - Personal Spending Behavioral Analysis API
  version: 1.0.0
  description: |
    API-first backend for the SIFT spending behavior analysis app.
    All frontend clients (web, future iOS) consume these endpoints.
    Deterministic computation precedes all AI usage.
    LLM cannot generate or modify numeric outputs.
servers:
  - url: /v1
    description: V1 API
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
    User:
      type: object
      required: [id, email]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        created_at:
          type: string
          format: date-time
...
    ImportJob:
      type: object
      required: [id, source, status]
      properties:
        id:
          type: string
          format: uuid
        source:
          type: string
          enum: [PLAID, CSV, PDF]
        status:
          type: string
          enum: [PENDING, PROCESSING, COMPLETED, FAILED]
        file_name:
          type: string
        row_count:
          type: integer
        error_count:
          type: integer
        errors:
          type: array
          items:
            type: object
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
...
    AnalyticsSignals:
      type: object
      properties:
        net_cashflow:
          type: number
        fixed_vs_variable_ratio:
          type: number
        required_vs_discretionary_ratio:
          type: number
        discretionary_burn_rate:
          type: number
        top_merchant_concentration:
          type: number
        category_creep_mom:
          type: number
        merchant_creep_mom:
          type: number
        rolling_baseline_delta:
          type: number
        seasonality_index:
          type: number
        price_drift_recurring:
          type: number
        day_of_week_map:
          type: object
          additionalProperties:
            type: number
        weekend_premium:
          type: number
        late_night_premium:
          type: number
        payday_spike:
          type: number
        impulse_frequency:
          type: number
        subscription_count:
          type: number
        subscription_total:
          type: number
        fees_total:
          type: number
        interest_total:
          type: number
        duplicate_charges:
          type: number
        cash_buffer_estimate:
          type: number
        volatility_score:
          type: number
        large_purchase_outliers:
          type: number
        savings_rate_estimate:
          type: number
        debt_pressure_proxy:
          type: number
    SavingsRecommendation:
      type: object
      required: [title, type, estimated_monthly_savings, estimated_annual_savings, confidence, proof, steps, score]
      properties:
        title:
          type: string
        type:
          type: string
          enum: [subscription, fee, interest, price_drift, habit_cut]
        estimated_monthly_savings:
          type: number
        estimated_annual_savings:
          type: number
        confidence:
          type: number
          minimum: 0
          maximum: 1
        proof:
          type: array
          items:
            type: object
            properties:
              txn_id:
                type: string
              description:
                type: string
              amount:
                type: number
              date:
                type: string
        steps:
          type: array
          items:
            type: string
        score:
          type: number
    Archetype:
      type: object
      required: [name, label, description]
      properties:
        name:
          type: string
          enum: [PAYDAY_SPIKER, WEEKEND_LEAKER, SUBSCRIPTION_CREEP, CONVENIENCE_EATER, IMPULSE_MICRO_SPENDER, LIFESTYLE_CREEP, BILL_DRIFT_VICTIM, FEE_PAYER, INTEREST_LEAKER, VOLATILITY_STRESSOR]
        label:
          type: string
        description:
          type: string
        triggered_by:
          type: array
          items:
            type: string
        signal_values:
          type: object
          additionalProperties:
            type: number
    AnalysisResult:
      type: object
      required: [signals, archetypes, recommendations, potential_annual_savings, period_start, period_end]
      properties:
        signals:
          $ref: '#/components/schemas/AnalyticsSignals'
        archetypes:
          type: array
          items:
            $ref: '#/components/schemas/Archetype'
          maxItems: 2
        recommendations:
          type: array
          items:
            $ref: '#/components/schemas/SavingsRecommendation'
          maxItems: 5
        potential_annual_savings:
          type: number
          description: Conservative with 0.85 safety haircut. Shown only if >= $1000.
        period_start:
          type: string
          format: date
        period_end:
          type: string
          format: date
        narration:
          type: object
          properties:
            summary:
              type: string
            archetype_narratives:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                  narrative:
                    type: string
            action_plans:
              type: array
              items:
                type: object
                properties:
                  title:
                    type: string
                  micro_plan:
                    type: string
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                name:
                  type: string
            example:
              email: user@example.com
              password: securepassword123
              name: Jane Doe
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
                      token:
                        type: string
              example:
                data:
                  user:
                    id: "550e8400-e29b-41d4-a716-446655440000"
                    email: user@example.com
                    name: Jane Doe
                    created_at: "2026-02-24T00:00:00Z"
                  token: "eyJhbGciOiJIUzI1NiIs..."
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
            example:
              email: user@example.com
              password: securepassword123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
                      token:
                        type: string
              example:
                data:
                  user:
                    id: "550e8400-e29b-41d4-a716-446655440000"
                    email: user@example.com
                    name: Jane Doe
                  token: "eyJhbGciOiJIUzI1NiIs..."
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /import/csv:
    post:
      tags: [Import]
      summary: Upload CSV statement
      description: |
        CSV must include date, amount, description columns.
        Idempotent by file hash.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, account_name]
              properties:
                file:
                  type: string
                  format: binary
                account_name:
                  type: string
                account_type:
                  type: string
                  default: checking
      responses:
        '200':
          description: Import started
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ImportJob'
              example:
                data:
                  id: "880e8400-e29b-41d4-a716-446655440003"
                  source: CSV
                  status: PROCESSING
                  file_name: "statements.csv"
        '400':
          description: Invalid CSV format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /import/{jobId}:
    get:
      tags: [Import]
      summary: Get import job status
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Import job details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ImportJob'
        '404':
          description: Job not found
...
  /analysis:
    post:
      tags: [Analysis]
      summary: Run full analysis pipeline
      description: |
        Triggers: normalization → classification → analytics → recommendations → archetypes → narration.
        Returns complete analysis result.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                include_narration:
                  type: boolean
                  default: true
      responses:
        '200':
          description: Analysis complete
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/AnalysisResult'
  /analysis/latest:
    get:
      tags: [Analysis]
      summary: Get most recent analysis snapshot
      responses:
        '200':
          description: Latest analysis
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/AnalysisResult'
        '404':
          description: No analysis found
  /analysis/history:
    get:
      tags: [Analysis]
      summary: List analysis snapshots
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Analysis history
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AnalysisResult'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
...